+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.8.1 (?????)
+
+### Fixed
+
+- Added `ext-pcre` to required deps
+
 ## 0.8.0 (2018-08-27)
 
 ### Added